The most significant changes in the code include the introduction of a new method of setting labels in `QdrantClusterController.cs`, the addition of a new method in `EntityHelper.cs`, and the addition of a `JsonConverter` attribute in `V1QdrantCollection.cs` and `VectorSpec.cs`. Additionally, the versions of `Neon.Operator` and `Neon.Operator.Xunit` were updated in `qdrant-operator.csproj` and `Test.qdrant-operator.csproj` respectively. A new test method was added in `TestUtils.cs` and a new file `VectorSpecConverter.cs` was created.
1. In `QdrantClusterController.cs`, the method of setting labels was changed from individual assignment to using the `SetLabels` method for various Kubernetes resources. This change was made to improve the efficiency of the code.
2. A new method `AddRange` was added to `EntityHelper.cs` to allow the addition of a collection of items to a source dictionary. This new method enhances the functionality of the code.
3. The `JsonConverter` attribute was added to the `VectorSpec` property in `V1QdrantCollection.cs` and to the `VectorSpec` class in `VectorSpec.cs`. This attribute allows for the customization of the serialization and deserialization of the `VectorSpec` class.
4. The versions of `Neon.Operator` and `Neon.Operator.Xunit` were updated in `qdrant-operator.csproj` and `Test.qdrant-operator.csproj` respectively. This update ensures that the project is using the latest versions of these dependencies.
5. A new test method `VectorSpecConverterTest` was added to `TestUtils.cs` to test the functionality of the `VectorSpec` class.
6. A new file `VectorSpecConverter.cs` was added. This file contains a `JsonConverter` for the `VectorSpec` class, which customizes the serialization and deserialization of the class.

+++ src/qdrant-operator/Controllers/QdrantCollectionController.cs
@@ -154,18 +154,21 @@ namespace QdrantOperator
 
             if (exists)
             {
-                var currentVectors = resource.Status?.CurrentSpec?.VectorSpec?.NamedVectors?.Select(nv => nv.Name).Distinct().Order();
-                var expectedVectors = resource.Spec.VectorSpec?.NamedVectors?.Select(nv => nv.Name).Distinct().Order();
-
-                if (!currentVectors.SequenceEqual(expectedVectors))
+                if (resource.Spec.VectorSpec?.NamedVectors != null)
                 {
-                    logger?.LogInformationEx(() => "Current vectors not equal to expected vectors, recreating collection. " +
-                    $"Current Vectors: {string.Join(',', currentVectors)}" +
-                    $"Expected Vectors {string.Join(',', expectedVectors)}");
+                    var currentVectors = resource.Status?.CurrentSpec?.VectorSpec?.NamedVectors?.Select(nv => nv.Name).Distinct().Order();
+                    var expectedVectors = resource.Spec.VectorSpec?.NamedVectors?.Select(nv => nv.Name).Distinct().Order();
 
-                    await qdrantClient.DeleteCollectionAsync(resource.Metadata.Name, TimeSpan.FromMinutes(5));
+                    if (!currentVectors.SequenceEqual(expectedVectors))
+                    {
+                        logger?.LogInformationEx(() => "Current vectors not equal to expected vectors, recreating collection. " +
+                        $"Current Vectors: {string.Join(',', currentVectors)}" +
+                        $"Expected Vectors {string.Join(',', expectedVectors)}");
 
-                    exists = false;
+                        await qdrantClient.DeleteCollectionAsync(resource.Metadata.Name, TimeSpan.FromMinutes(5));
+
+                        exists = false;
+                    }
                 }
             }
